refactor: :truck: Rename yet again, welcome Nularc
Alongside the rename, I also took the time to improve the structure of the documentation, and improve the documentation site's responsiveness and features. I added the `docsify-pagination` plugin, so you can go to the next/previous page easily.

`Using` statements in existing projects will break, though since probably no one except me uses this, it's not that big of a deal.